requirements.txt
selenium>=4.15.0
pandas>=2.1.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
pyyaml>=6.0
xlrd>=2.0.0